Time for a refresh!
Some config title strings are now pretty.
We can do this, since we're no longer limited by the awful old shell
based system, but just use JSON.

- - -

Better colors in curses.
It looked bad in Termux, and especially my new theme in KDE.

- - -

Retired plugins (some were already retired):
- auto-pairs
- vim-numbertoggle
- lightline
- feline
- vim-gitgutter
- vim-rainbow

Added plugins (half to replace older ones):
- telescope
- nvim-autopairs
- nvim-numbertoggle
- indent-blankline

- - -

Attempted to fix ufo config and turn it back on.
Still not sure it helps at all with the folding itself, I feel like
treesitter does all the work, but we've got some new clickable things in
the left column, so we'll see if we want to keep those... At some point
I'll work on getting those moved to the right side of the line numbers,
and the git gutter stuff on the left (as it already is).

- - -

Updated treesitter.
Got tired of all the random auto indentation failures, discovered
treesitter can do it far better than whatever did it before. Also turned
on special highlighting for the heck of it.